+++ modules/creat_book.js
@@ -1,6 +1,8 @@
+/* eslint-disable import/prefer-default-export */
+
 export class StandardBook {
-    constructor(title, author) {
-      this.title = title;
-      this.author = author;
-    }
+  constructor(title, author) {
+    this.title = title;
+    this.author = author;
+  }
 }
