+++ modules/get_local_storage.js
@@ -1,5 +1,7 @@
-export let getData = () => {
+/* eslint-disable import/prefer-default-export */
+
+export const getData = () => {
   const localdata = localStorage.getItem('localdata');
   const dataStored = JSON.parse(localdata);
   return dataStored;
-}
+};
