+++ modules/update_local_storage.js
@@ -1,4 +1,6 @@
-export let updateLocalStorage = (obj) => {
-    const localdata = JSON.stringify(obj);
-    localStorage.setItem('localdata', localdata);
-}
+/* eslint-disable import/prefer-default-export */
+
+export const updateLocalStorage = (obj) => {
+  const localdata = JSON.stringify(obj);
+  localStorage.setItem('localdata', localdata);
+};
